… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/testdata/upstream/at_modifier.test
… pkg/streamingpromql/testdata/upstream/functions.test
… pkg/streamingpromql/testdata/upstream/native_histograms.test
… pkg/streamingpromql/testdata/upstream/subquery.test
… pkg/streamingpromql/testing.go
… pkg/streamingpromql/types/data.go
… pkg/streamingpromql/types/fpoint_ring_buffer.go
… pkg/streamingpromql/types/hpoint_ring_buffer.go
… pkg/streamingpromql/types/operator.go
… pkg/streamingpromql/types/ring_buffer_test.go