@charleskorn 2024-10-21
>In doing so, it would also be nice to see some more nested sub-queries (eg 3+ sub-queries deep). Additionally different offsets or @ etc in the sub-queries would be good to check for correct step alignment.

There are already a bunch of tests that cover offsets, @ and nesting (albeit not three levels deep) in the upstream test cases. Is there something in particular that you'd like to see covered that's not already covered there?