@charleskorn 2024-10-21
>Lastly, perhaps also a test, or a once off benchmark, for where we check at the extremes of having points at the end of the buffer outside the range (ie to check the previously discussed performance regression).

Not sure I follow this sorry - this case is exercised for every range query with more than one step, and is covered by those benchmarks as well.