@jhesketh 2024-10-21
> > In doing so, it would also be nice to see some more nested sub-queries (eg 3+ sub-queries deep). Additionally different offsets or @ etc in the sub-queries would be good to check for correct step alignment.
> 
> There are already a bunch of tests that cover offsets, @ and nesting (albeit not three levels deep) in the upstream test cases. Is there something in particular that you'd like to see covered that's not already covered there?

Where the step continues to move from the parent query at each depth level. (even if this doesn't really happen due to aligning with epoch).



> > Lastly, perhaps also a test, or a once off benchmark, for where we check at the extremes of having points at the end of the buffer outside the range (ie to check the previously discussed performance regression).
> 
> Not sure I follow this sorry - this case is exercised for every range query with more than one step, and is covered by those benchmarks as well.

I was thinking as a one-off benchmark that does it at an extreme, but you're probably right that this isn't necessary.