@charleskorn 2024-10-21
> Where the step continues to move from the parent query at each depth level. (even if this doesn't really happen due to aligning with epoch).

Done, I've added a test like this in 220e5de.